Add Integration tests
