+++ blog_os/Cargo.toml
@@ -24,3 +24,7 @@ features = ["spin_no_std"]
 test-args = ["-device", "isa-debug-exit,iobase=0xf4,iosize=0x04", "-serial", "stdio", "-display", "none"]
 test-success-exit-code = 33         # (0x10 << 1) | 1
 test-timeout = 300  # Seconds
+
+[[test]]
+name = "should_panic"
+harness = false
