+++ blog_os/src/main.rs
@@ -1,14 +1,14 @@
 #![no_std]
 #![no_main]
 #![feature(custom_test_frameworks)]
-#![test_runner(crate::test_runner)]
+#![test_runner(blog_os::test_runner)]
 #![reexport_test_harness_main = "test_main"]
 
-mod vga_buffer;
-mod serial;
 use core::panic::PanicInfo;
+use blog_os::println;
 
-/// This func is called onPanic by compiler
+
+// This func is called onPanic by compiler
 #[cfg(not(test))]
 #[panic_handler]
 fn panic(info: &PanicInfo) -> !{
@@ -19,31 +19,15 @@ fn panic(info: &PanicInfo) -> !{
 #[cfg(test)]
 #[panic_handler]
 fn panic(info: &PanicInfo) -> ! {
-    serial_println!("[failed]\n");
-    serial_println!("Error: {}\n", info);
-    exit_qemu(QemuExitCode::Failed);
-    loop {}
+    blog_os::test_panic_handler(info)
 }
 
-static HELLO: &[u8] = b"Red rust is a mood!";
+// static HELLO: &[u8] = b"Red rust is a mood!";
 
 #[no_mangle]
 pub extern "C" fn _start() -> ! {
     
-    // vga_buffer::print_something();
-    
-    use core::fmt::Write;
-    //vga_buffer::WRITER.lock().write_str("Hello from the other side!!").unwrap();
-    //write!(vga_buffer::WRITER.lock(), ", some numbers: {} {}", 42, 1.337).unwrap();
     println!("Red Rust is {}!!", "eternal");
-    /*let vga_buffer = 0xb8000 as *mut u8;
-
-    for (i, &byte) in HELLO.iter().enumerate() {
-        unsafe {
-            *vga_buffer.offset(i as isize * 2) = byte;
-            *vga_buffer.offset(i as isize * 2 + 1) = 0xb;
-        }
-    }*/
     
     #[cfg(test)]
     test_main();
@@ -52,15 +36,6 @@ pub extern "C" fn _start() -> ! {
 }
 
 
-#[cfg(test)]
-pub fn test_runner(tests: &[&dyn Testable]) {
-    serial_println!("Running {} tests", tests.len());
-    for test in tests {
-        test.run();
-    }
-    exit_qemu(QemuExitCode::Success);
-}
-
 #[test_case]
 fn trivial_assertion() {
     // serial_print!("trivial assertion... ");
@@ -68,33 +43,4 @@ fn trivial_assertion() {
     // serial_println!("[ok]");
 }
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-#[repr(u32)]
-pub enum QemuExitCode {
-    Success = 0x10,
-    Failed = 0x11,
-}
 
-pub fn exit_qemu(exit_code: QemuExitCode) {
-    use x86_64::instructions::port::Port;
-
-    unsafe {
-        let mut port = Port::new(0xf4);
-        port.write(exit_code as u32);
-    }
-}
-
-pub trait Testable {
-    fn run(&self) -> ();
-}
-
-impl<T> Testable for T
-where
-    T: Fn(),
-{
-    fn run(&self) {
-        serial_print!("{}...\t", core::any::type_name::<T>());
-        self();
-        serial_println!("[ok]");
-    }
-}
