+++ blog_os/tests/basic_boot.rs
@@ -2,25 +2,31 @@
 #![no_main]
 
 #![feature(custom_test_frameworks)]
-#![test_runner(crate::test_runner)]
+#![test_runner(blog_os::test_runner)]
 #![reexport_test_harness_main = "test_main"]
 
 use core::panic::PanicInfo;
+use blog_os::println;
+use blog_os::test_panic_handler;
 
 #[no_mangle]
 pub extern "C" fn _start() -> ! {
     test_main();
-
     loop {}
 }
 
 
-fn test_runner(tests: &[&dyn Fn()]) {
-    unimplemented!();
-}
-
+//fn test_runner(tests: &[&dyn Fn()]) {
+//    unimplemented!();
+//}
+//
 
 #[panic_handler]
 fn panic(info: &PanicInfo) -> ! {
-    loop {}
+    blog_os::test_panic_handler(info)
+}
+
+#[test_case]
+fn test_println() {
+    println!("test_println output");
 }
